requirements.txt
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
importlib-metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
zipp==3.21.0
psutil==6.1.1
numpy==2.2.1
pillow==11.0.0
spidev==3.6
gpiozero==2.0.1
git-lfs==1.6
python-dotenv==1.0.1
schedule==1.2.2
timer-event==0.9.1
watchdog==6.0.0